add api upgrade ticket

+++ services/order-service/src/main/java/com/example/cronjob/DTO/Response/TicketResponse.java
@@ -12,7 +12,7 @@ public record TicketResponse(
         Long ticketTypeId,
         String name,
         String ticketCode,
-        float actualPrice,
+        double actualPrice,
         LocalDateTime validFrom,
         LocalDateTime validUntil,
         TicketStatus status,
